@@ -16,17 +16,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.2.3",
-    "@angular/cdk": "^17.2.2",
-    "@angular/common": "^17.2.3",
-    "@angular/compiler": "^17.2.3",
-    "@angular/core": "^17.2.3",
-    "@angular/forms": "^17.2.3",
-    "@angular/localize": "^17.2.3",
-    "@angular/material": "^17.2.2",
-    "@angular/platform-browser": "^17.2.3",
-    "@angular/platform-browser-dynamic": "^17.2.3",
-    "@angular/router": "^17.2.3",
+    "@angular/animations": "^19.2.14",
+    "@angular/cdk": "^19.2.14",
+    "@angular/common": "^19.2.14",
+    "@angular/compiler": "^19.2.14",
+    "@angular/core": "^19.2.14",
+    "@angular/forms": "^19.2.14",
+    "@angular/localize": "^19.2.14",
+    "@angular/material": "^19.2.14",
+    "@angular/platform-browser": "^19.2.14",
+    "@angular/platform-browser-dynamic": "^19.2.14",
+    "@angular/router": "^19.2.14",
     "@badeball/cypress-cucumber-preprocessor": "^22.0.1",
     "@bahmutov/cypress-esbuild-preprocessor": "^2.2.0",
     "@coreui/icons-angular": "^5.4.13",
@@ -94,34 +94,34 @@
     "@storybook/manager-webpack5": "^6.5.16",
     "@storybook/test-runner": "^0.22.0",
     "@storybook/testing-library": "~0.2.2",
-    "@types/jest": "29.5.12",
-    "@types/node": "20.11.25",
-    "@typescript-eslint/eslint-plugin": "7.1.1",
-    "@typescript-eslint/parser": "7.1.1",
-    "babel-loader": "^9.1.3",
-    "chromatic": "^11.25.2",
-    "cypress": "^13.6.6",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^20.17.57",
+    "@typescript-eslint/eslint-plugin": "^7.18.0",
+    "@typescript-eslint/parser": "^8.33.0",
+    "babel-loader": "^10.0.0",
+    "chromatic": "^12.0.0",
+    "cypress": "^14.4.0",
     "cypress-cucumber-preprocessor": "^4.3.1",
     "eslint": "~8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-cypress": "^2.15.1",
-    "eslint-plugin-storybook": "^0.8.0",
-    "jest": "29.7.0",
-    "jest-environment-jsdom": "^29.4.1",
-    "jest-preset-angular": "^14.5.3",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-cypress": "^5.0.1",
+    "eslint-plugin-storybook": "^9.0.4",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^30.0.0-beta.3",
+    "jest-preset-angular": "^14.6.0",
     "jsonc-eslint-parser": "^2.4.0",
     "mini-css-extract-plugin": "^2.8.1",
-    "nx": "^18.0.7",
-    "nx-cloud": "^19.0.0",
-    "prettier": "^3.2.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "sass": "^1.71.1",
-    "sass-loader": "^14.1.1",
+    "nx": "^19.8.4",
+    "nx-cloud": "^19.1.0",
+    "prettier": "^3.5.3",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "sass": "^1.89.1",
+    "sass-loader": "^16.0.5",
     "storybook": "^7.6.6",
-    "ts-jest": "29.1.2",
+    "ts-jest": "29.3.4",
     "ts-node": "^10.9.2",
-    "typescript": "5.3.3",
+    "typescript": "5.5.4",
     "webpack": "^5.95.0"
   }
 }
