@@ -1,19 +1,20 @@
-import { CommonModule } from '@angular/common';
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 
 @Component({
   standalone: true,
   // Needs to be a different name to the CLI template button
   selector: 'lda-storybook-button',
-  imports: [CommonModule],
+  imports: [],
   template: `
     <button
       type="button"
       (click)="onClick.emit($event)"
-      [ngClass]="classes"
-      [ngStyle]="{ 'background-color': backgroundColor }"
+      [class]="classes"
+      [style]="{ 'background-color': backgroundColor }"
     >
-      <i *ngIf="icon" class="ti-arrow-left"></i>
+      @if(icon){
+      <i class="ti-arrow-left"></i>
+      }
       {{ label }}
     </button>`,
   styleUrls: ['./button.css'],
