@@ -2,12 +2,11 @@ import { Component } from '@angular/core';
 import {selectTechBlogById, selectTechLoading} from "../../+state/pages.selector";
 import {Store} from "@ngrx/store";
 import {BlogDetailComponent} from "./blog-detail/blog-detail.component";
-import {NgIf} from "@angular/common";
 
 @Component({
   standalone: true,
   selector: 'lda-blog-details',
-  imports: [BlogDetailComponent, NgIf],
+  imports: [BlogDetailComponent],
   template: `
     @if (!loading()){
       <lda-blog-detail
