@@ -6,13 +6,12 @@ import {selectTechDesc, selectTechErrorMessage, selectTechLoading} from "../+sta
 import {PagesStore} from "../pages.store";
 import {Page404Component} from "../page404/page404.component";
 import {BlogListComponent} from "./blog-list/blog-list.component";
-import {SpinnerComponent} from "../../../../taskbox/src/spinner/spinner.component";
 
 @UntilDestroy()
 @Component({
   standalone: true,
   selector: 'lda-blog',
-  imports: [Page404Component, BlogListComponent, SpinnerComponent],
+  imports: [Page404Component, BlogListComponent],
   template: `
       @if (!loading() && errorMessage() === '') {
       <div
@@ -63,7 +62,6 @@ import {SpinnerComponent} from "../../../../taskbox/src/spinner/spinner.componen
         </div>
       </div>
       } @else if(loading()) {
-        <lda-spinner [isLoading]="true" [message]="'Please wait...'"></lda-spinner>
         <div>Please wait...</div>
       } @else if (errorMessage()){
         <!-- Error Message -->
