@@ -1,11 +1,10 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { ButtonModule } from '../button';
 import PageComponent from "../pages/page.component";
 import HeaderComponent from "../header/header.component";
 
 @NgModule({
-  imports: [CommonModule, ButtonModule, PageComponent, HeaderComponent],
+  imports: [ButtonModule, PageComponent, HeaderComponent],
   exports: [
     ButtonModule,
     PageComponent,
